tests/wpt/tests/css/css-backgrounds/parsing/border-image-repeat-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Backgrounds and Borders: getComputedStyle().borderImageRepeat</title>
<link rel="help" href="https://drafts.csswg.org/css-backgrounds/#border-image-repeat">
<meta name="assert" content="border-image-repeat computed value is specified keywords.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("border-image-repeat", "round");
test_computed_value("border-image-repeat", "stretch repeat");
test_computed_value("border-image-repeat", "round space");
</script>
</body>
</html>
